==> notebooks/06_visualization.ipynb <==
%load_ext autoreload
%autoreload 2

import sys
import os

# Add the parent directory to the Python path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

import pandas as pd

from src.config import TRANSFORMED_DATA_DIR

features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")

features = features_and_targets.iloc[:, :-1]
targets = features_and_targets["target"]


features[features['pickup_location_id']=='HB101']

from src.plot_utils import plot_aggregated_time_series

plot_aggregated_time_series(features, targets, pickup_location_id='HB101')
